+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `TagType::remove_from` now works for ID3v2 tags in APE and FLAC files
   - This previously verified that the `FileType` supported the tag. It now has special exceptions for these formats to
     allow stripping out these unsupported tags
+- **MP4**: Renamed `AdvisoryRating::None` to `AdvisoryRating::Inoffensive`
 
 ### Fixed
 - **MP4**: Non-full `meta` atoms are now properly handled.